se implemento el metodo de interseccion de paredes, testeado en el commit pasado, tambien se opto por cambiar las coordenadas del punto de int a float, para hacerlo mas generico y por las intersecciones

+++ Obligatorio1_Arancet_Cohen/Logic.Test/PointTest.cs
@@ -18,14 +18,14 @@ namespace Logic.Test
         [TestMethod]
         public void GetCoordXTest(){
             int expectedResult = 3;
-            int actualResult= instance.CoordX;
+            float actualResult= instance.CoordX;
             Assert.AreEqual(expectedResult,actualResult);
         }
 
         [TestMethod]
         public void GetCoordYTest(){
             int expectedResult = 2;
-            int actualResult = instance.CoordY;
+            float actualResult = instance.CoordY;
             Assert.AreEqual(expectedResult, actualResult);
         }
